refactor

+++ TumblrReader/TBLAPIManager.h
@@ -1,5 +1,5 @@
 //
-//  TBLServiceManager.h
+//  TBLAPIManager.h
 //  TumblrReader
 //
 //  Created by Adam on 10.06.2016.
@@ -21,19 +21,18 @@
 
 + (nonnull TBLAPIManager *)sharedManager;
 
-- (void)
-fetchPostsForUsername:(NSString *_Nonnull)username
-       startPostIndex:(int)startPostIndex
-           postsCount:(int)postsCount
-              success:(void (^_Nonnull)(NSURLSessionTask *_Nonnull task,
-                                        TBLBlogMeta *_Nullable blogMeta,
-                                        NSArray<TBLPost *> *_Nullable posts,
-                                        NSError *_Nullable error))success
-              failure:(void (^_Nonnull)(NSURLSessionTask *_Nullable task,
-                                        NSError *_Nonnull error))failure;
+- (void)fetchPostsForUsername:(NSString *_Nonnull)username
+               startPostIndex:(NSUInteger)startPostIndex
+                   postsCount:(NSUInteger)postsCount
+                      success:(void (^ _Nonnull)(NSURLSessionTask *_Nonnull task,
+                              TBLBlogMeta *_Nullable blogMeta,
+                              NSArray<TBLPost *> *_Nullable posts,
+                              NSError *_Nullable error))success
+                      failure:(void (^ _Nonnull)(NSURLSessionTask *_Nullable task,
+                              NSError *_Nonnull error))failure;
 
 - (void)imageFromURLString:(NSString *_Nonnull)URLString
-                   success:(void (^_Nonnull)(UIImage *_Nullable image))success
-                   failure:(void (^_Nonnull)(NSError *_Nonnull error))failure;
+                   success:(void (^ _Nonnull)(UIImage *_Nullable image))success
+                   failure:(void (^ _Nonnull)(NSError *_Nonnull error))failure;
 
 @end
